+++ tests/bootstrap_test.go
@@ -13,9 +13,9 @@ import (
 	"testing"
 	"time"
 
-	"github.com/EasyPost/easypost-go/v3"
 	"github.com/dnaeon/go-vcr/cassette"
 	"github.com/dnaeon/go-vcr/recorder"
+	"github.com/elmarw/easypost-go/v3"
 	"github.com/stretchr/testify/suite"
 )