+++ tests/easypost_test.go
@@ -1,7 +1,7 @@
 package easypost_test
 
 import (
-	"github.com/EasyPost/easypost-go/v3"
+	"github.com/elmarw/easypost-go/v3"
 )
 
 func (c *ClientTests) TestClientTimeout() {